Changed pos from double to int

+++ GameProject/Engine/Components/FreeMove.h
@@ -17,7 +17,7 @@ private:
 	float speed;
 	float sensitivity;
 	bool mouseLock;
-	double xPos, yPos;
+	int xPos, yPos;
 	std::unordered_map<unsigned, bool> pressedKeys;
 
 	void moveKeyboard(KeyEvent * evnt);
